Naive solutions to both Day 01 problems
Problem_01a simply reads the input file and keeps a running total.

Problem_01b keeps a cache of the previous sums and compares each time.

A smarter solution to 01b would be to keep the cache sorted and use
binary search and insertion.
This is equivalent to using an appropriate tree structure.