Initial solution to first Day 04 problem
This was a lot more effort than I thought.
I cheated by using system commands to sort the timestamped events, and
to count the number of nights.
Each night stores a guard ID, and a bit-field of asleep for each minute.
After reading that in, determine the unique guards, and process the
saved information accordingly.

Very fiddly.