Aggressive optimisation for both Day 05 problems
Within the confines of the problem specification, checking for the two
adjacent characters being the same letter of opposite case reduces to
merely checking that the difference between them is equal to the
ASCII offset between the banks of upper and lower case.

Shaves off about a third of the execution time for both parts.