Implement stack-based solution to the Day 05 problems
Represent the polymer as a stack (linked-list).
Compare adjacent letters by popping from a full stack to an empty one.
Push when the letters are different, skip when they match.

Drastically reduces the amount of comparison necessary and utterly
slashes the runtime.

A future commit will make this more pronounced, once the routines for
each Day are merged, giving 2 results and the respective timestamps.