Solved both Day 09 problems using a cyclical doubly-linked list
The hardest part was getting the %Remove() pointer-juggling to work.
I wasn't fooled by the 100* factor - just use INT64.

The apparent run-time discrepancies are due to the %Destroy() calls at
the end of the solving routine.
It may be possible to handle this more efficiently, but the timing of
these should be handled explicitly.