Implement solution to both Day 11 problems using a summed-area table
The solution to the first part is an early intermediate result while
solving the second part, which the timing logic reflects.

There was some difficulty involved in getting the summed-area table
working, due to the various offsets and array-bound juggling required to
be useful for an inclusive sum over the input element indices.

Also fixed a typo in one of the Makefiles.