Solve the second Day 07 problem
This was a natural extension of the first part, though some effort was
required to reason out how to allocate tasks to workers.

A cleaner solution might have methods on Type(Worker) which treat them
like their own queue.
Removing the nesting of DO loops and IF statements would make the
actual approach much clearer.